have done some change in Readme (#20)
"Expectation is your extension only works when the user is on GitHub and can interact with the page. This interaction is to add a button that links to the user's profile."
to 
"The expectation is your extension only works when the user is on GitHub and can interact with the page. This interaction adds a button that links to the user's profile."